Do not calculate MatchResults eagerly
There are many variables that should be taken into account when
calculating the MatchResult, so it did not make sense to calculate
them all at once at the start of the League. Also the user could change
players, and other variables during the League that could affect future
results. Even though this is not how it currently works, this is the
first step to make sure we can take those variables into account.