User Club squad rating to calculate MatchResult
This uses `ClubSquad.rating()` to calculate a team rating for the
squad that can be used to affect the odds of winning/losing a
match. The rating takes into account the squad mean rating and
standard deviation and uses a normal distribution to sample a
random rating.
This is then fed into the `MatchResultGenerator` to aid on the
random result calculation. The rating difference between home
and away is used to calculate the odds.
Also created `ClubDataStore` to help easily query data related
to clubs from `ElifutDataStore`.