Be smarter about building the squads
No longer require at least 4 defenders and 4 midfielders. We now
just require a goalkeeper + 10 players and just fill them up into
positions as best as we can. We can put an attacker as defender if
needed to complete the formation.
Also improves the app initialization routine by skipping clubs
with less than 11 players and skipping players that are not of
regular quality.